+++ src/tests/fixtures/mpd_client.rs
@@ -519,6 +519,10 @@ impl MpdClient for TestMpdClient {
     fn move_in_queue(&mut self, _from: SingleOrRange, _to: QueueMoveTarget) -> MpdResult<()> {
         todo!("Not yet implemented")
     }
+
+    fn delete_from_queue(&mut self, _songs: SingleOrRange) -> MpdResult<()> {
+        todo!("Not yet implemented")
+    }
 }
 
 impl SocketClient for TestMpdClient {